fraudgraph/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
neo4j==5.14.1
confluent-kafka==2.3.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
reportlab==4.0.7
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
